Server tracking/query to own app gameservers
Unsure how much I want to put into this app, I need to find a good way to separate Network Rules from server rules, as I think Network rules falls more into the main app/CMS than gameservers

Effectively just ported over functionality to new app from raptormc. No major noteworthy changes to the code as of yet besides whitespace reduction.
