+++ raptorWeb/raptormc/urls.py
@@ -21,9 +21,5 @@ urlpatterns = [
     path('user/', SR.User_Views.SiteMembers.as_view(), name='site_members'),
     path('user/<str:profile_name>', SR.User_Views.User_Page.as_view(), name="user_page"),
     path('user/<str:profile_name>/edit', SR.User_Views.User_Edit_Page.as_view(), name="user_edit_page"),
-    # Ajax
-    path('server_button_poll/', SR.Ajax_Views.Server_Buttons.as_view(), name="server_buttons_poll"),
-    path('server_modal_poll/', SR.Ajax_Views.Server_Modals.as_view(), name="server_modals_poll"),
-    path('total_count_poll/', SR.Ajax_Views.Total_Count.as_view(), name="totol_counts_poll")
 
 ]
