+++ raptorWeb/raptormc/views.py
@@ -1,11 +1,8 @@
-from django.shortcuts import render
 from django.views.generic import TemplateView
-from django.http import HttpResponseRedirect
 from os.path import join
 from logging import getLogger
 
 from raptorWeb import settings
-from raptormc.jobs import player_poller, update_context
 from raptormc.util import viewContext
 
 TEMPLATE_DIR_RAPTORMC = join(settings.TEMPLATE_DIR, "raptormc")
@@ -29,7 +26,7 @@ class ShadowRaptor():
 
             def get_context_data(self, **kwargs): 
                 context = super().get_context_data(**kwargs)
-                context.update(player_poller.currentPlayers_DB)
+                
                 return viewContext.update_context(context = context, informative_text_names = ["Homepage Information"], announcements=True)
 
         class Announcements(TemplateView):
@@ -40,7 +37,7 @@ class ShadowRaptor():
 
             def get_context_data(self, **kwargs):
                 context = super().get_context_data(**kwargs)
-                context.update(player_poller.currentPlayers_DB)
+                
                 return viewContext.update_context(context = context, informative_text_names = ["Announcements Information"], announcements=True)
 
         class Rules(TemplateView):
@@ -51,7 +48,7 @@ class ShadowRaptor():
 
             def get_context_data(self, **kwargs):
                 context = super().get_context_data(**kwargs)
-                context.update(player_poller.currentPlayers_DB)
+                
                 return viewContext.update_context(context = context, informative_text_names = [
                     "Rules Information",
                     "Network Rules"])
@@ -64,7 +61,7 @@ class ShadowRaptor():
 
             def get_context_data(self, **kwargs):
                 context = super().get_context_data(**kwargs)
-                context.update(player_poller.currentPlayers_DB)
+                
                 return viewContext.update_context(context = context, informative_text_names = ["Banned Items Information"])
 
         class Voting(TemplateView):
@@ -75,7 +72,7 @@ class ShadowRaptor():
 
             def get_context_data(self, **kwargs):
                 context = super().get_context_data(**kwargs)
-                context.update(player_poller.currentPlayers_DB)
+                
                 return viewContext.update_context(context = context, informative_text_names = ["Voting Information"])
 
         class HowToJoin(TemplateView):
@@ -86,7 +83,7 @@ class ShadowRaptor():
 
             def get_context_data(self, **kwargs):
                 context = super().get_context_data(**kwargs)
-                context.update(player_poller.currentPlayers_DB)
+                
                 return viewContext.update_context(context = context, informative_text_names = [
                     "Joining Information",
                     "Using the CurseForge Launcher",
@@ -101,7 +98,7 @@ class ShadowRaptor():
 
             def get_context_data(self, **kwargs):
                 context = super().get_context_data(**kwargs)
-                context.update(player_poller.currentPlayers_DB)
+                
                 return viewContext.update_context(context = context, informative_text_names = ["Staff App Information"])
 
     class User_Views():
@@ -116,7 +113,7 @@ class ShadowRaptor():
 
             def get_context_data(self, **kwargs):
                 context = super().get_context_data(**kwargs)
-                context.update(player_poller.currentPlayers_DB)
+                
                 return viewContext.update_context(context = context)
 
         class User_Page(TemplateView):
@@ -127,7 +124,7 @@ class ShadowRaptor():
 
             def get_context_data(self, **kwargs):
                 context = super().get_context_data(**kwargs)
-                context.update(player_poller.currentPlayers_DB)
+                
                 return viewContext.update_context(context = context)
 
         class User_Edit_Page(TemplateView):
@@ -138,45 +135,5 @@ class ShadowRaptor():
 
             def get_context_data(self, **kwargs):
                 context = super().get_context_data(**kwargs)
-                context.update(player_poller.currentPlayers_DB)
+                
                 return viewContext.update_context(context = context)
-
-    class Ajax_Views():
-        """
-        Views that return HTML for use in AJAX requests
-        """
-        class Server_Buttons(TemplateView):
-            """
-            Returns HTML of server buttons
-            Attempts to fetch new info before sending
-            """
-            def get(self, request):
-                if request.headers.get('HX-Request') == "true":
-                    template_name = join(settings.RAPTOMC_TEMPLATE_DIR, 'serverButtons.html')
-                    return render(request, template_name, context=player_poller.currentPlayers_DB)
-                else:
-                    return HttpResponseRedirect('../')
-
-        class Server_Modals(TemplateView):
-            """
-            Returns HTML of server modals, as well as player counts modal
-            Attempts to fetch new info before sending
-            """
-            def get(self, request):
-                if request.headers.get('HX-Request') == "true":
-                    template_name = join(settings.RAPTOMC_TEMPLATE_DIR, 'serverModals.html')
-                    return render(request, template_name, context=player_poller.currentPlayers_DB)
-                else:
-                    return HttpResponseRedirect('../')
-
-        class Total_Count(TemplateView):
-            """
-            Returns a simple HttpResponse with the total count of players on all servers
-            """
-            def get(self, request):
-                if request.headers.get('HX-Request') == "true":
-                    template_name = join(settings.RAPTOMC_TEMPLATE_DIR, 'playerCounts.html')
-                    update_context()
-                    return render(request, template_name, context=player_poller.currentPlayers_DB)
-                else:
-                    return HttpResponseRedirect('../')
