+++ raptorWeb/templates/gameservers/playerCounts.html
@@ -24,7 +24,7 @@
     </div>
 </div>
 
-<button class="mt-auto btn btn-lg btn-dark" data-bs-toggle="modal" data-bs-target="#playerNames" id="totalPlayers">
+<button class="mt-auto btn btn-lg btn-dark mb-3" data-bs-toggle="modal" data-bs-target="#playerNames" id="totalPlayers">
     <aside class="fs-4 text-white opacity-75 p-3 pt-0 pb-0" >Currently Playing: <span id="total_count" class="fs-1">
     {{ totalCount }}</span></aside>
     <aside class="fs-6 text-white opacity-75 p-3 pt-0 pb-0">Click to view players!</aside>
