+++ raptorWeb/templates/gameservers/serverButtons.html
@@ -6,7 +6,7 @@
     {% if server == 'view' or server == 'totalCount' %}
     {% else %}
                     
-    <span id="serverButton" class="fade-in-button col-lg-12 col-6 mt-2"><button class="btn btn-dark fs-4 d-flex" data-bs-toggle="modal" data-bs-target="#{{ server|get_key }}Info">
+    <span id="serverButton" class="fade-in-button col-lg-12 col-6 mt-2"><button class="btn btn-dark fs-4 d-flex mt-2" data-bs-toggle="modal" data-bs-target="#{{ server|get_key }}Info">
         <img class="w-100 fade-in" src="/media/modpack_pictures/{{ server|get_key}}.webp" alt="{{ server|get_modpack_name }}">
         {% if server|get_maintenance %}
         <img class="indicatorImage p-2 fade-in" src="{% static 'image/maintenance_indicator.webp' %}">
