+++ raptorWeb/templates/gameservers/serverButtons_loading.html
@@ -1,0 +1,15 @@
+{% load static %}
+
+<div hx-get="/servers/server_button_poll" hx-trigger="load delay:0.3s" hx-target="#server_button" hx-swap="innerHTML"></div>
+
+<div id="server_button">
+
+{% for server in server_info reversed %}
+    <span id="serverButton" class="col-lg-12 col-6 mt-2 opacity-50">
+    <button class="btn btn-dark fs-4 w-100 align-content-center">
+        <img id="loadingSpinnerButtons" src="{% static 'image/three-dots.svg' %}" alt="Loading..." class="align-middle opacity-75">
+    </button>
+    </span>
+{% endfor %}
+
+</div>
